requirements.txt
numpy==1.16.4
Pillow==6.0.0
joblib==0.13.2
